requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
bcrypt==4.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.4.4
cryptography==42.0.5
ecdsa==0.19.1
executing==2.0.1
fastapi==0.110.0
h11==0.14.0
icecream==2.1.3
idna==3.6
mailersend==0.5.6
Mako==1.3.2
MarkupSafe==2.1.5
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
PyMySQL==1.1.0
pypng==0.20220715.0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
qrcode==7.4.2
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
